requirements.txt
# =============================================
# JIRA REPORT SYSTEM — ЗАВИСИМОСТИ
# =============================================
# Python: 3.7.3+ (Debian 10 buster)
# Все версии проверены на совместимость с Python 3.7
# =============================================

# Core (веб-фреймворк и кэш)
Flask==2.0.3
Flask-Caching==2.0.2

# Обработка данных
pandas==1.3.5
openpyxl==3.1.2

# Jira API
jira==3.2.0

# Утилиты
python-dotenv==0.21.1
python-dateutil==2.8.2
requests==2.31.0
urllib3==1.26.18
tenacity==8.2.3

# Web server
gunicorn==20.1.0

# =============================================
# Database (для истории отчётов)
# SQLAlchemy 2.0+ требует Python 3.7+
# =============================================
SQLAlchemy==2.0.23
alembic==1.12.1

# =============================================
# PDF экспорт
# weasyprint 52.5 — последняя версия с полной поддержкой Python 3.7
# =============================================
weasyprint==52.5

# =============================================
# Планировщик задач
# APScheduler 3.10.0+ требует Python 3.8+
# Используем версию 3.9.1 для Python 3.7
# =============================================
APScheduler==3.9.1.post1

# =============================================
# Telegram бот
# python-telegram-bot 20.0+ требует Python 3.8+
# Используем версию 13.15 для Python 3.7
# =============================================
python-telegram-bot==13.15

# =============================================
# Визуализация (дополнительно к Chart.js)
# plotly 5.x поддерживает Python 3.7
# =============================================
plotly==5.18.0

# =============================================
# Development dependencies
# Версии совместимые с Python 3.7
# =============================================
flake8==5.0.4
black==22.12.0
mypy==0.991
pytest==7.4.3
pytest-cov==4.1.0
